# Disable result-cache in LevelsTestCase