The Rule Levels tests are pretty slow. its the slowest part of the CI pipline at.

Before this PR, it takes about 11m30s to run Rule Levels in CI.

Instead of writing and clearing the result cache, use `--debug` instead which prevents usage of the result cache

----

I have reduced the test locally to short the feedback loop for the change:

After the PR:
```
➜  phpstan-src git:(1.10.x) ✗ time make tests-levels
php vendor/bin/paratest --runner WrapperRunner --no-coverage --group levels
ParaTest v6.6.3 upon PHPUnit 9.5.23 #StandWithUkraine

..                                                                  2 / 2 (100%)

Time: 00:31.752, Memory: 44.00 MB

OK (2 tests, 52 assertions)
make tests-levels  28.26s user 3.87s system 101% cpu 31.788 total          ##  After
```

Before the PR
```
➜  phpstan-src git:(debug-levels) ✗ time make tests-levels
php vendor/bin/paratest --runner WrapperRunner --no-coverage --group levels
ParaTest v6.6.3 upon PHPUnit 9.5.23 #StandWithUkraine

..                                                                  2 / 2 (100%)

Time: 00:39.511, Memory: 54.00 MB

OK (2 tests, 52 assertions)
make tests-levels  32.79s user 6.68s system 99% cpu 39.562 total    ## Before
```


I think its now ~30% faster