@staabm 2023-05-20
hmm interessting.. the local speedup does not reproduce in CI